+++ website/documentation-site.ts
@@ -21,6 +21,7 @@ export const DOCUMENTATION_DEFAULT_VERSION = 'latest';
 
 export const DOCUMENTATION_VERSIONS = [
   'latest',
+  '0.47.0',
   '0.46.0',
   '0.45.0',
   '0.44.0',